GH-327 + GH-266 Handle not found and duplicate entities
* Handle Duplicate Streams and Tasks with dedicated Error Handlers for
  - DuplicateStreamDefinitionException
  - DuplicateTaskException
* Properly report not found entities as 404 and cover
  - NoSuchStreamDefinitionException
  - NoSuchTaskDefinitionException
  - MetricsMvcEndpoint.NoSuchMetricException
* Properly report failure to register a module/library

This fixes spring-cloud/spring-cloud-dataflow#266
This fixes spring-cloud/spring-cloud-dataflow#327

+++ spring-cloud-dataflow-admin-starter/src/main/java/org/springframework/cloud/dataflow/admin/controller/StreamDefinitionController.java
@@ -25,6 +25,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.cloud.dataflow.admin.repository.DuplicateStreamDefinitionException;
+import org.springframework.cloud.dataflow.admin.repository.NoSuchStreamDefinitionException;
 import org.springframework.cloud.dataflow.admin.repository.StreamDefinitionRepository;
 import org.springframework.cloud.dataflow.core.ModuleDefinition;
 import org.springframework.cloud.dataflow.core.ModuleDeploymentId;
@@ -131,12 +132,6 @@ public class StreamDefinitionController {
 	                 @RequestParam("definition") String dsl,
 	                 @RequestParam(value = "deploy", defaultValue = "true")
 	                 boolean deploy) {
-		if (this.repository.exists(name)) {
-			throw new DuplicateStreamDefinitionException(
-					String.format("Cannot create stream %s because another one has already " +
-							"been created with the same name", name));
-		}
-
 		StreamDefinition stream = new StreamDefinition(name, dsl);
 		this.repository.save(stream);
 		if (deploy) {
@@ -151,6 +146,9 @@ public class StreamDefinitionController {
 	@RequestMapping(value = "/{name}", method = RequestMethod.DELETE)
 	@ResponseStatus(HttpStatus.OK)
 	public void delete(@PathVariable("name") String name) {
+		if(repository.findOne(name) == null) {
+			throw new NoSuchStreamDefinitionException(name);
+		}
 		deploymentController.undeploy(name);
 		this.repository.delete(name);
 	}
@@ -162,7 +160,11 @@ public class StreamDefinitionController {
 	@RequestMapping(value = "/{name}", method = RequestMethod.GET)
 	@ResponseStatus(HttpStatus.OK)
 	public StreamDefinitionResource display(@PathVariable("name") String name) {
-		return streamDefinitionAssembler.toResource(repository.findOne(name));
+		StreamDefinition definition = repository.findOne(name);
+		if (definition == null) {
+			throw new NoSuchStreamDefinitionException(name);
+		}
+		return streamDefinitionAssembler.toResource(definition);
 	}
 
 	/**
